USPS Routes improvements
Instead of returning the sometimes inaccurate city listed in the EDDM database (see sketch's post in #mapomatic-scripts on 8/13), the script now uses the ZIP code returned for each route to look up the USPS recommend and other recognized city names. The returned routes may also be clicked to view the recommended and other recognized cities in a WW alert.